% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{analyze_outbreaks}
\alias{analyze_outbreaks}
\title{Analyze outbreak characteristics}
\usage{
analyze_outbreaks(results)
}
\arguments{
\item{results}{Simulation results}
}
\value{
Data frame with outbreak statistics
}
\description{
Analyze outbreak characteristics
}
