% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{animate_spatial_spread}
\alias{animate_spatial_spread}
\title{Create animated spatial plot over time}
\usage{
animate_spatial_spread(
  results,
  compartment = "I",
  replicate = 1,
  n_rows = 5,
  n_cols = 5,
  time_points = NULL
)
}
\arguments{
\item{results}{plague_results object (must be spatial)}

\item{compartment}{Compartment to animate (default "I")}

\item{replicate}{Which replicate to plot (default 1)}

\item{n_rows}{Number of rows in spatial grid (default 5)}

\item{n_cols}{Number of columns in spatial grid (default 5)}

\item{time_points}{Vector of time points to include (NULL for all)}
}
\value{
gganimate object
}
\description{
Create animated spatial plot over time
}
