% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{briere_T_opt}
\alias{briere_T_opt}
\title{Temperature at which a Briere response peaks}
\usage{
briere_T_opt(T_min, T_max, q)
}
\arguments{
\item{T_min, T_max, q}{Briere parameters.}
}
\value{
The optimum temperature.
}
\description{
Convenience for converting Briere coordinates to the optimum, e.g. when
summarising posterior draws.
}
