% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calculate_force_of_infection}
\alias{calculate_force_of_infection}
\title{Calculate force of infection over time}
\usage{
calculate_force_of_infection(results)
}
\arguments{
\item{results}{plague_results object}
}
\value{
Tibble with force of infection by time
}
\description{
Calculate force of infection over time
}
