% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{calculate_intervention_effects}
\alias{calculate_intervention_effects}
\title{Calculate intervention effects}
\usage{
calculate_intervention_effects(scenario_results)
}
\arguments{
\item{scenario_results}{Results from multiple scenarios}
}
\value{
Data frame with intervention statistics
}
\description{
Calculate intervention effects
}
