% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{calculate_model_stats}
\alias{calculate_model_stats}
\title{Calculate model statistics}
\usage{
calculate_model_stats(output)
}
\arguments{
\item{output}{Simulation output}
}
\value{
List of statistics
}
\description{
Calculate model statistics
}
