% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calculate_R0}
\alias{calculate_R0}
\title{Calculate basic reproduction number (R0) for different model types}
\usage{
calculate_R0(params, model_type = "rats_only")
}
\arguments{
\item{params}{plague_parameters object or list}

\item{model_type}{Type of model ("rats_only", "with_humans", etc.)}
}
\value{
Numeric R0 value or list of R0 values
}
\description{
Calculate basic reproduction number (R0) for different model types
}
