% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_deterministic.R
\name{create_deterministic_human_model}
\alias{create_deterministic_human_model}
\title{Create deterministic plague model with human dynamics}
\usage{
create_deterministic_human_model()
}
\value{
odin model generator
}
\description{
Create deterministic plague model with human dynamics
}
