% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_deterministic.R
\name{create_deterministic_model}
\alias{create_deterministic_model}
\title{Create base deterministic plague model (rats and fleas only)}
\usage{
create_deterministic_model()
}
\value{
odin model generator
}
\description{
Create base deterministic plague model (rats and fleas only)
}
