% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_stochastic.R
\name{create_human_stochastic_model}
\alias{create_human_stochastic_model}
\title{Create human stochastic plague model}
\usage{
create_human_stochastic_model()
}
\value{
odin.dust model generator
}
\description{
Create human stochastic plague model
}
