% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_deterministic.R
\name{create_seasonal_deterministic_model}
\alias{create_seasonal_deterministic_model}
\title{Create deterministic plague model with seasonal forcing}
\usage{
create_seasonal_deterministic_model()
}
\value{
odin model generator
}
\description{
Create deterministic plague model with seasonal forcing
}
