% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{create_seasonal_forcing}
\alias{create_seasonal_forcing}
\title{Create seasonal forcing}
\usage{
create_seasonal_forcing(times, amplitude = 0.2)
}
\arguments{
\item{times}{Vector of timepoints}

\item{amplitude}{Seasonal amplitude}
}
\value{
Vector of seasonal multipliers
}
\description{
Create seasonal forcing
}
