% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_stochastic.R
\name{create_spatial_stochastic_model}
\alias{create_spatial_stochastic_model}
\title{Create spatial stochastic plague model}
\usage{
create_spatial_stochastic_model()
}
\value{
odin.dust model generator
}
\description{
Create spatial stochastic plague model
}
