% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{create_timepoints}
\alias{create_timepoints}
\title{Create simulation timepoints}
\usage{
create_timepoints(years = 40, timestep = 1/365)
}
\arguments{
\item{years}{Number of years to simulate}

\item{timestep}{Time step in years}
}
\value{
Vector of timepoints
}
\description{
Create simulation timepoints
}
