% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_mod_diag_strip.R
\name{diag_strip_server}
\alias{diag_strip_server}
\title{Diagnostics strip module — server.}
\usage{
diag_strip_server(id, lab_session)
}
\arguments{
\item{id}{Module namespace id.}

\item{lab_session}{A \link{LabSession} instance.}
}
\value{
Invisibly, the moduleServer result.
}
\description{
Diagnostics strip module — server.
}
