% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_gadget.R
\name{explore_plague_model}
\alias{explore_plague_model}
\title{Create a Shiny gadget for exploring the plague model}
\usage{
explore_plague_model(initial_params = NULL)
}
\arguments{
\item{initial_params}{Optional list of initial parameters}
}
\value{
Modified parameters if Apply is clicked, NULL if Cancel
}
\description{
Create a Shiny gadget for exploring the plague model
}
