% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{format_parameters}
\alias{format_parameters}
\title{Format parameters for display}
\usage{
format_parameters(params)
}
\arguments{
\item{params}{Parameter list}
}
\value{
Character string with formatted parameters
}
\description{
Format parameters for display
}
