% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{get_grid_positions}
\alias{get_grid_positions}
\title{Convert population indices to grid coordinates}
\usage{
get_grid_positions(indices, n_cols)
}
\arguments{
\item{indices}{Population indices}

\item{n_cols}{Number of columns}
}
\value{
Data frame with row and column positions
}
\description{
Convert population indices to grid coordinates
}
