% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{get_seasonal_forcing}
\alias{get_seasonal_forcing}
\title{Generate seasonal forcing}
\usage{
get_seasonal_forcing(timesteps, amplitude = 0.2)
}
\arguments{
\item{timesteps}{Vector of timesteps}

\item{amplitude}{Amplitude of seasonal forcing}
}
\value{
Vector of seasonal multipliers
}
\description{
Generate seasonal forcing
}
