% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_mod_hero.R
\name{hero_ui}
\alias{hero_ui}
\title{Hero plot module — UI.}
\usage{
hero_ui(id)
}
\arguments{
\item{id}{Module namespace id.}
}
\value{
A \code{bslib::card()} with the hero plot.
}
\description{
Hero plot module — UI.
}
