% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{load_named_parameters}
\alias{load_named_parameters}
\title{Load named parameter sets}
\usage{
load_named_parameters(name)
}
\arguments{
\item{name}{Name of parameter set}
}
\value{
List of parameters
}
\description{
Load named parameter sets
}
