% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{load_named_scenario}
\alias{load_named_scenario}
\title{Load named scenario parameters}
\usage{
load_named_scenario(name)
}
\arguments{
\item{name}{Name of scenario}
}
\value{
List of model parameters
}
\description{
Load named scenario parameters
}
