% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{load_parameters}
\alias{load_parameters}
\title{Load plague model parameters}
\usage{
load_parameters(param_set = "defaults", validate = TRUE, ...)
}
\arguments{
\item{param_set}{Name of parameter set or path to YAML file or list of parameters}

\item{validate}{Logical, whether to validate parameters}

\item{...}{Additional parameters to override}
}
\value{
List of validated parameters
}
\description{
Load plague model parameters
}
