% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{load_scenario}
\alias{load_scenario}
\title{Load plague model scenario parameters}
\usage{
load_scenario(scenario = "defaults", validate = TRUE, ...)
}
\arguments{
\item{scenario}{Name of scenario or path to YAML file or list of parameters}

\item{validate}{Logical, whether to validate parameters}

\item{...}{Additional model parameters to override}
}
\value{
List of validated model parameters
}
\description{
Load plague model scenario parameters
}
