% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{make_contact_matrix}
\alias{make_contact_matrix}
\title{Create contact matrix for grid-based metapopulation}
\usage{
make_contact_matrix(n_rows = 5, n_cols = 5)
}
\arguments{
\item{n_rows}{Number of rows in grid}

\item{n_cols}{Number of columns in grid}
}
\value{
Normalized contact matrix
}
\description{
Create contact matrix for grid-based metapopulation
}
