% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_mod_model.R
\name{model_accordion_server}
\alias{model_accordion_server}
\title{Model accordion module — server.}
\usage{
model_accordion_server(id, lab_session)
}
\arguments{
\item{id}{Module namespace id.}

\item{lab_session}{A \link{LabSession} instance.}
}
\value{
Invisibly, the moduleServer result.
}
\description{
Model accordion module — server.
}
