% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_mod_model.R
\name{model_accordion_ui}
\alias{model_accordion_ui}
\title{Model accordion module — UI.}
\usage{
model_accordion_ui(id)
}
\arguments{
\item{id}{Module namespace id.}
}
\value{
A \code{bslib::accordion()} containing the model config controls.
}
\description{
Model accordion module — UI.
}
