% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{new_plague_results}
\alias{new_plague_results}
\title{Create a plague_results object}
\usage{
new_plague_results(data, model_type, params, run_info = list())
}
\arguments{
\item{data}{Tibble with simulation results}

\item{model_type}{Type of model used}

\item{params}{Parameters used in simulation}

\item{run_info}{Runtime information}
}
\value{
plague_results object
}
\description{
Create a plague_results object
}
