% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{plague_stochastic_humans}
\alias{plague_stochastic_humans}
\title{Stochastic humans plague model (carcass formulation)}
\description{
An odin2/dust2 system generator compiled at package build time from
\code{inst/odin/plague_stochastic_humans.R}. Pass it directly to
\code{\link[dust2:dust_system_create]{dust2::dust_system_create()}} or \code{\link[dust2:dust_filter_create]{dust2::dust_filter_create()}}.
}
