% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Create multi-panel comparison plot}
\usage{
plot_comparison(results_list, compartment = "I", population = 1)
}
\arguments{
\item{results_list}{Named list of plague_results objects}

\item{compartment}{Compartment to compare (default "I")}

\item{population}{Which population to plot (default 1)}
}
\value{
ggplot2 object
}
\description{
Create multi-panel comparison plot
}
