% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{plot_connectivity}
\alias{plot_connectivity}
\title{Plot connectivity matrix}
\usage{
plot_connectivity(contact_matrix, n_rows, n_cols)
}
\arguments{
\item{contact_matrix}{Contact matrix}

\item{n_rows}{Number of rows}

\item{n_cols}{Number of columns}
}
\description{
Plot connectivity matrix
}
