% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dynamics}
\alias{plot_dynamics}
\title{Plot compartment dynamics with uncertainty bands}
\usage{
plot_dynamics(
  results,
  compartments = NULL,
  population = 1,
  show_uncertainty = TRUE,
  log_scale = FALSE
)
}
\arguments{
\item{results}{plague_results object}

\item{compartments}{Vector of compartments to plot (NULL for all)}

\item{population}{Which population to plot (for spatial models, default 1)}

\item{show_uncertainty}{Show uncertainty bands for stochastic models (default TRUE)}

\item{log_scale}{Use log scale for y-axis (default FALSE)}
}
\value{
ggplot2 object
}
\description{
Plot compartment dynamics with uncertainty bands
}
