% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{plot_outbreak_distribution}
\alias{plot_outbreak_distribution}
\title{Plot outbreak size distribution}
\usage{
plot_outbreak_distribution(outbreak_stats)
}
\arguments{
\item{outbreak_stats}{Output from analyze_outbreaks}
}
\value{
ggplot object
}
\description{
Plot outbreak size distribution
}
