% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{plot_patch_dynamics}
\alias{plot_patch_dynamics}
\title{Plot dynamics for specific patches}
\usage{
plot_patch_dynamics(results, patches, compartments)
}
\arguments{
\item{results}{Simulation results}

\item{patches}{Vector of patch indices to plot}

\item{compartments}{Vector of compartments to plot}
}
\value{
ggplot object
}
\description{
Plot dynamics for specific patches
}
