% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_phase_portrait}
\alias{plot_phase_portrait}
\title{Plot phase portrait (S vs I)}
\usage{
plot_phase_portrait(
  results,
  compartments = c("S_r", "I_r"),
  population = 1,
  replicate = 1,
  ...
)
}
\arguments{
\item{results}{plague_results object}

\item{compartments}{Vector of two compartments to plot (default c("S_r", "I_r"))}

\item{population}{Which population to plot (for spatial models, default 1)}

\item{replicate}{Which replicate to plot (for stochastic models, default 1)}

\item{...}{Additional ggplot2 arguments}
}
\value{
ggplot2 object
}
\description{
Plot phase portrait (S vs I)
}
