% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Plot parameter sensitivity results}
\usage{
plot_sensitivity(sensitivity_results, compartment = "I", metric = "peak")
}
\arguments{
\item{sensitivity_results}{Output from run_sensitivity_analysis}

\item{compartment}{Compartment to plot (default "I")}

\item{metric}{Summary metric to plot ("peak", "final", "auc")}
}
\value{
ggplot2 object
}
\description{
Plot parameter sensitivity results
}
