% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_spatial_heatmap}
\alias{plot_spatial_heatmap}
\title{Plot spatial heatmap for a single time point}
\usage{
plot_spatial_heatmap(
  results,
  time_point,
  compartment = "I",
  replicate = 1,
  n_rows = 5,
  n_cols = 5
)
}
\arguments{
\item{results}{plague_results object (must be spatial)}

\item{time_point}{Time point to plot}

\item{compartment}{Compartment to plot (default "I")}

\item{replicate}{Which replicate to plot (default 1)}

\item{n_rows}{Number of rows in spatial grid (default 5)}

\item{n_cols}{Number of columns in spatial grid (default 5)}
}
\value{
ggplot2 object
}
\description{
Plot spatial heatmap for a single time point
}
