% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{plot_total_infected}
\alias{plot_total_infected}
\title{Plot total infected over time with uncertainty}
\usage{
plot_total_infected(results)
}
\arguments{
\item{results}{Simulation results}
}
\value{
ggplot object
}
\description{
Plot total infected over time with uncertainty
}
