% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{print.plague_outbreak_metrics}
\alias{print.plague_outbreak_metrics}
\title{Print method for outbreak metrics}
\usage{
\method{print}{plague_outbreak_metrics}(x, ...)
}
\arguments{
\item{x}{plague_outbreak_metrics object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for outbreak metrics
}
