% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{print.plague_results}
\alias{print.plague_results}
\title{Print method for plague_results}
\usage{
\method{print}{plague_results}(x, ...)
}
\arguments{
\item{x}{plague_results object}

\item{...}{Additional arguments passed to tibble print}
}
\description{
Print method for plague_results
}
