% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{print.scenario_parameters}
\alias{print.scenario_parameters}
\title{Print method for plague_parameters}
\usage{
\method{print}{scenario_parameters}(x, ...)
}
\arguments{
\item{x}{plague_parameters object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for plague_parameters
}
