% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_mod_priors.R
\name{priors_accordion_server}
\alias{priors_accordion_server}
\title{Priors accordion module — server.}
\usage{
priors_accordion_server(id, lab_session)
}
\arguments{
\item{id}{Module namespace id.}

\item{lab_session}{A \link{LabSession} instance.}
}
\value{
Invisibly, the moduleServer result.
}
\description{
Priors accordion module — server.
}
