% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_mod_priors.R
\name{priors_accordion_ui}
\alias{priors_accordion_ui}
\title{Priors accordion module — UI.}
\usage{
priors_accordion_ui(id)
}
\arguments{
\item{id}{Module namespace id.}
}
\value{
A \code{bslib::card()} whose body is the dynamic priors accordion.
}
\description{
Priors accordion module — UI.
}
