% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_priors.R
\name{priors_to_dsl}
\alias{priors_to_dsl}
\title{Convert a full priors list to a list of monty-DSL expressions.}
\usage{
priors_to_dsl(priors)
}
\arguments{
\item{priors}{Named list of priors (see file header).}
}
\value{
List of unevaluated expressions in the same order as \code{priors}.
}
\description{
Convert a full priors list to a list of monty-DSL expressions.
}
