% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{run_deterministic_model}
\alias{run_deterministic_model}
\title{Run deterministic model}
\usage{
run_deterministic_model(params, times, include_humans, seasonal)
}
\arguments{
\item{params}{List of parameters}

\item{times}{Vector of time points}

\item{include_humans}{Include human dynamics}

\item{seasonal}{Include seasonal forcing}
}
\value{
Tidy tibble with results
}
\description{
Run deterministic model
}
