% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{run_plague_simulation}
\alias{run_plague_simulation}
\title{Run plague model simulation}
\usage{
run_plague_simulation(
  params,
  times = NULL,
  seasonal = FALSE,
  seasonal_amplitude = 0.2
)
}
\arguments{
\item{params}{List of model parameters}

\item{times}{Vector of timepoints}

\item{seasonal}{Logical, whether to include seasonal forcing}

\item{seasonal_amplitude}{Amplitude of seasonal forcing if used}
}
\value{
Tibble with simulation results
}
\description{
Run plague model simulation
}
