% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{run_sensitivity_analysis}
\alias{run_sensitivity_analysis}
\title{Create simulation timepoints}
\usage{
run_sensitivity_analysis(
  base_params,
  param_name,
  range = seq(0.5, 1.5, by = 0.1),
  seasonal = FALSE
)
}
\arguments{
\item{base_params}{Base parameter list}

\item{param_name}{Parameter to vary}

\item{range}{Vector of multipliers}

\item{seasonal}{Include seasonal forcing}

\item{years}{Number of years to simulate}

\item{timestep}{Time step in years}

\item{output}{Simulation output tibble}

\item{log_scale}{Logical, whether to use log scale}

\item{plot_type}{One of "all", "infected", or "phase"}
}
\value{
Vector of timepoints
Plot plague simulation results

ggplot object
Run sensitivity analysis

Tibble with sensitivity analysis results
}
\description{
Create simulation timepoints
}
