% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{run_spatial_stochastic_model}
\alias{run_spatial_stochastic_model}
\title{Run spatial stochastic model}
\usage{
run_spatial_stochastic_model(params, timesteps, n_particles, n_threads)
}
\arguments{
\item{params}{List of parameters}

\item{timesteps}{Vector of timesteps}

\item{n_particles}{Number of particles}

\item{n_threads}{Number of threads}
}
\value{
Tidy tibble with results
}
\description{
Run spatial stochastic model
}
