% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{run_stochastic_simulation}
\alias{run_stochastic_simulation}
\title{Run stochastic simulation with given parameters}
\usage{
run_stochastic_simulation(params, timesteps, n_particles = 1, n_threads = 1)
}
\arguments{
\item{params}{List of parameters}

\item{timesteps}{Vector of timesteps}

\item{n_particles}{Number of particles (replicates)}

\item{n_threads}{Number of threads for parallel processing}
}
\value{
Data frame with simulation results
}
\description{
Run stochastic simulation with given parameters
}
