% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{save_parameters}
\alias{save_parameters}
\title{Save parameter set to YAML}
\usage{
save_parameters(params, file)
}
\arguments{
\item{params}{Parameter list}

\item{file}{Filename to save to}
}
\value{
Invisibly returns the parameter list
}
\description{
Save parameter set to YAML
}
