% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_mod_status.R
\name{status_bar_ui}
\alias{status_bar_ui}
\title{Status bar module — UI.}
\usage{
status_bar_ui(id)
}
\arguments{
\item{id}{Module namespace id.}
}
\value{
A \code{shiny::div()} styled as a sticky bottom bar.
}
\description{
Status bar module — UI.
}
