% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarize_outbreak_metrics}
\alias{summarize_outbreak_metrics}
\title{Summarize outbreak metrics across replicates}
\usage{
summarize_outbreak_metrics(outbreak_metrics)
}
\arguments{
\item{outbreak_metrics}{Output from calculate_outbreak_metrics}
}
\value{
Summary statistics
}
\description{
Summarize outbreak metrics across replicates
}
