% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{summary.plague_results}
\alias{summary.plague_results}
\title{Summary method for plague_results}
\usage{
\method{summary}{plague_results}(object, ...)
}
\arguments{
\item{object}{plague_results object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for plague_results
}
