% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate plague model parameters}
\usage{
validate_parameters(params)
}
\arguments{
\item{params}{List of parameters}
}
\value{
TRUE if valid, stops with error if invalid
}
\description{
Validate plague model parameters
}
