% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yersinia-package.R
\docType{package}
\name{yersinia-package}
\alias{yersinia}
\alias{yersinia-package}
\title{yersinia: Plague Transmission Models for Epidemiological Research}
\description{
A toolkit for plague transmission modeling using stochastic compartmental models. Implements carcass-based transmission dynamics (Didelot et al. 2017) with spatial structure and multi-host dynamics for epidemiological research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flmnh-ai/yersinia}
  \item \url{https://flmnh-ai.github.io/yersinia/}
  \item Report bugs at \url{https://github.com/flmnh-ai/yersinia/issues}
}

}
\author{
\strong{Maintainer}: Nick Gauthier \email{nicolas.gauthier@ufl.edu} (\href{https://orcid.org/0000-0002-2225-5827}{ORCID})

Authors:
\itemize{
  \item Nich Martin \email{n.martin@ufl.edu} (\href{https://orcid.org/0000-0001-9539-250X}{ORCID})
}

}
\keyword{internal}
