% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calculate_equilibrium}
\alias{calculate_equilibrium}
\title{Calculate equilibrium values for deterministic models}
\usage{
calculate_equilibrium(params, model_type = "rats_only")
}
\arguments{
\item{params}{plague_parameters object or list}

\item{model_type}{Type of model}
}
\value{
List of equilibrium values
}
\description{
Calculate equilibrium values for deterministic models
}
\note{
This function is for reference/educational purposes.
For operational models, use the stochastic simulations in run_plague_model().
}
