% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_mod_cohort.R
\name{cohort_ui}
\alias{cohort_ui}
\title{Cohort module — UI.}
\usage{
cohort_ui(id)
}
\arguments{
\item{id}{Module namespace id.}
}
\value{
A \code{bslib::card()} with the cohort chips + picker grid.
}
\description{
Cohort module — UI.
}
