% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{make_contact_matrix}
\alias{make_contact_matrix}
\title{Calculate model statistics}
\usage{
make_contact_matrix(n_rows = 5, n_cols = 5)
}
\arguments{
\item{n_rows}{Number of rows in grid}

\item{n_cols}{Number of columns in grid}

\item{output}{Simulation output}
}
\value{
List of statistics
Create contact matrix for grid-based metapopulation

Normalized contact matrix
}
\description{
Calculate model statistics
}
