% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{plot.plague_results}
\alias{plot.plague_results}
\title{Plot method for plague_results}
\usage{
\method{plot}{plague_results}(x, compartments = NULL, log_scale = FALSE, ...)
}
\arguments{
\item{x}{plague_results object}

\item{compartments}{Vector of compartments to plot (NULL for all)}

\item{log_scale}{Logical, use log scale for y-axis}

\item{...}{Additional arguments (ignored)}
}
\description{
Plot method for plague_results
}
