% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R
\name{plot_plague_simulation}
\alias{plot_plague_simulation}
\title{Create simulation timepoints}
\usage{
plot_plague_simulation(output, log_scale = FALSE, plot_type = "all")
}
\arguments{
\item{output}{Simulation output tibble}

\item{log_scale}{Logical, whether to use log scale}

\item{plot_type}{One of "all", "infected", or "phase"}

\item{years}{Number of years to simulate}

\item{timestep}{Time step in years}
}
\value{
Vector of timepoints
Plot plague simulation results

ggplot object
}
\description{
Create simulation timepoints
}
