% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_utils.R, R/plotting.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Plot sensitivity analysis results}
\usage{
plot_sensitivity(sensitivity_results, compartment = "I_r", metric = "peak")

plot_sensitivity(sensitivity_results, compartment = "I_r", metric = "peak")
}
\arguments{
\item{sensitivity_results}{Output from run_sensitivity_analysis}

\item{compartment}{Compartment to plot (default "I_r")}

\item{metric}{Summary metric to plot ("peak", "final", "auc")}

\item{variable}{Variable to plot}
}
\value{
ggplot object

ggplot2 object
}
\description{
Plot sensitivity analysis results

Plot parameter sensitivity results
}
