% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_mod_status.R
\name{status_bar_server}
\alias{status_bar_server}
\title{Status bar module — server.}
\usage{
status_bar_server(id, lab_session)
}
\arguments{
\item{id}{Module namespace id.}

\item{lab_session}{A \link{LabSession} instance.}
}
\value{
Invisibly, the moduleServer result.
}
\description{
Status bar module — server.
}
